should get font directly from font manager. not by game host.
also should register this manager.
